+++ sympy/printing/lambdarepr.py
@@ -11,6 +11,7 @@ from sympy.core.basic import Basic
 from sympy.external import import_module
 from sympy.utilities import default_sort_key
 
+
 class LambdaPrinter(PythonCodePrinter):
     """
     This printer converts expressions into strings that can be used by
@@ -57,227 +58,6 @@ class LambdaPrinter(PythonCodePrinter):
 
     def _print_NumberSymbol(self, expr):
         return str(expr)
-
-
-class TensorflowPrinter(PythonCodePrinter):
-    """
-    Tensorflow printer which handles vectorized piecewise functions,
-    logical operators, max/min, and relational operators.
-    """
-    printmethod = "_tensorflowcode"
-
-    def _print_And(self, expr):
-        "Logical And printer"
-        # We have to override the printer because it uses Python 'and' keyword.
-        # If the printer didn't define it, we could use StrPrinter's
-        # version of the function and add 'logical_and' to TENSORFLOW_TRANSLATIONS.
-        return '{0}({1})'.format('logical_and', ','.join(self._print(i) for i in expr.args))
-
-    def _print_Or(self, expr):
-        "Logical Or printer"
-        # We have to override the printer because it uses Python 'or' keyword.
-        # If the printer didn't define it, we could use StrPrinter's
-        # version of the function and add 'logical_or' to TENSORFLOW_TRANSLATIONS.
-        return '{0}({1})'.format('logical_or', ','.join(self._print(i) for i in expr.args))
-
-    def _print_Not(self, expr):
-        "Logical Not printer"
-        # We have to override the printer because it uses Python 'not' keyword.
-        # If the printer didn't define it, we would still have to define our
-        #     own because StrPrinter doesn't define it.
-        return '{0}({1})'.format('logical_not', ','.join(self._print(i) for i in expr.args))
-
-    def _print_Min(self, expr):
-        from sympy import Min
-        if len(expr.args) == 1:
-            return self._print(expr.args[0])
-
-        return 'minimum({0}, {1})'.format(
-            self._print(expr.args[0]),
-            self._print(Min(*expr.args[1:])))
-
-    def _print_Max(self, expr):
-        from sympy import Max
-        if len(expr.args) == 1:
-            return self._print(expr.args[0])
-
-        return 'maximum({0}, {1})'.format(
-            self._print(expr.args[0]),
-            self._print(Max(*expr.args[1:])))
-
-    def _print_Piecewise(self, expr):
-        tensorflow = import_module('tensorflow')
-        if tensorflow and V(tensorflow.__version__) < '1.0':
-            tensorflow_piecewise = "select"
-        else:
-            tensorflow_piecewise = "where"
-
-        from sympy import Piecewise
-        e, cond = expr.args[0].args
-        if len(expr.args) == 1:
-            return '{0}({1}, {2}, {3})'.format(
-                tensorflow_piecewise,
-                self._print(cond),
-                self._print(e),
-                0)
-
-        return '{0}({1}, {2}, {3})'.format(
-            tensorflow_piecewise,
-            self._print(cond),
-            self._print(e),
-            self._print(Piecewise(*expr.args[1:])))
-
-    def _print_Relational(self, expr):
-        "Relational printer for Equality and Unequality"
-        op = {
-            '==' :'equal',
-            '!=' :'not_equal',
-            '<'  :'less',
-            '<=' :'less_equal',
-            '>'  :'greater',
-            '>=' :'greater_equal',
-        }
-        if expr.rel_op in op:
-            lhs = self._print(expr.lhs)
-            rhs = self._print(expr.rhs)
-            return '{op}({lhs}, {rhs})'.format(op=op[expr.rel_op],
-                                               lhs=lhs,
-                                               rhs=rhs)
-        return super(TensorflowPrinter, self)._print_Relational(expr)
-
-    def _print_MatrixBase(self, expr):
-        return "%s(%s)" % (
-            self._module_format("tensorflow.constant"),
-            str(expr.tolist()),
-        )
-
-    def _print_MatMul(self, expr):
-        return self._expand_fold_binary_op("tensorflow.matmul", expr.args)
-
-    def _print_Assignment(self, expr):
-        # TODO: is this necessary?
-        return "%s = %s" % (
-            self._print(expr.lhs),
-            self._print(expr.rhs),
-        )
-
-    def _print_CodeBlock(self, expr):
-        # TODO: is this necessary?
-        ret = []
-        for subexpr in expr.args:
-            ret.append(self._print(subexpr))
-        return "\n".join(ret)
-
-    def _get_letter_generator_for_einsum(self):
-        for i in range(97, 123):
-            yield chr(i)
-        for i in range(65, 91):
-            yield chr(i)
-        raise ValueError("out of letters")
-
-    def _print_CodegenArrayTensorProduct(self, expr):
-        array_list = [j for i, arg in enumerate(expr.args) for j in
-                (self._print(arg), "[%i, %i]" % (2*i, 2*i+1))]
-        letters = self._get_letter_generator_for_einsum()
-        contraction_string = ",".join(["".join([next(letters) for j in range(i)]) for i in expr.subranks])
-        return '%s("%s", %s)' % (
-                self._module_format('tensorflow.einsum'),
-                contraction_string,
-                ", ".join([self._print(arg) for arg in expr.args])
-        )
-
-    def _get_einsum_string(self, subranks, contraction_indices):
-        letters = self._get_letter_generator_for_einsum()
-        contraction_string = ""
-        # TODO: share with NumPy
-        counter = 0
-        d = {j: min(i) for i in contraction_indices for j in i}
-        indices = []
-        for rank_arg in subranks:
-            lindices = []
-            for i in range(rank_arg):
-                if counter in d:
-                    lindices.append(d[counter])
-                else:
-                    lindices.append(counter)
-                counter += 1
-            indices.append(lindices)
-        mapping = {}
-        letters_free = []
-        letters_dum = []
-        for i in indices:
-            for j in i:
-                if j not in mapping:
-                    l = next(letters)
-                    mapping[j] = l
-                else:
-                    l = mapping[j]
-                contraction_string += l
-                if j in d:
-                    if l not in letters_dum:
-                        letters_dum.append(l)
-                else:
-                    letters_free.append(l)
-            contraction_string += ","
-        contraction_string = contraction_string[:-1]
-        return contraction_string, letters_free, letters_dum
-
-    def _print_CodegenArrayContraction(self, expr):
-        from sympy.codegen.array_utils import CodegenArrayTensorProduct
-        base = expr.expr
-        contraction_indices = expr.contraction_indices
-        contraction_string, letters_free, letters_dum = self._get_einsum_string(base.subranks, contraction_indices)
-
-        if len(contraction_indices) == 0:
-            return self._print(base)
-        if isinstance(base, CodegenArrayTensorProduct):
-            elems = ["%s" % (self._print(arg)) for arg in base.args]
-            return "%s(\"%s\", %s)" % (
-                self._module_format("tensorflow.einsum"),
-                contraction_string,
-                ", ".join(elems)
-            )
-        raise NotImplementedError()
-
-    def _print_CodegenArrayDiagonal(self, expr):
-        from sympy.codegen.array_utils import CodegenArrayTensorProduct
-        diagonal_indices = list(expr.diagonal_indices)
-        if len(diagonal_indices) > 1:
-            # TODO: this should be handled in sympy.codegen.array_utils,
-            # possibly by creating the possibility of unfolding the
-            # CodegenArrayDiagonal object into nested ones. Same reasoning for
-            # the array contraction.
-            raise NotImplementedError
-        if len(diagonal_indices[0]) != 2:
-            raise NotImplementedError
-        if isinstance(expr.expr, CodegenArrayTensorProduct):
-            subranks = expr.expr.subranks
-            elems = expr.expr.args
-        else:
-            subranks = expr.subranks
-            elems = [expr.expr]
-        diagonal_string, letters_free, letters_dum = self._get_einsum_string(subranks, diagonal_indices)
-        elems = [self._print(i) for i in elems]
-        return '%s("%s", %s)' % (
-            self._module_format("tensorflow.einsum"),
-            "{0}->{1}{2}".format(diagonal_string, "".join(letters_free), "".join(letters_dum)),
-            ", ".join(elems)
-        )
-
-    def _print_CodegenArrayPermuteDims(self, expr):
-        return "%s(%s, %s)" % (
-            self._module_format("tensorflow.transpose"),
-            self._print(expr.expr),
-            self._print(expr.permutation.args[0]),
-        )
-
-    def _print_CodegenArrayElementwiseAdd(self, expr):
-        return self._expand_fold_binary_op('tensorflow.add', expr.args)
-
-
-def tensorflow_code(expr):
-    printer = TensorflowPrinter()
-    return printer.doprint(expr)
 
 
 # numexpr works by altering the string passed to numexpr.evaluate
