+++ sympy/printing/tests/test_tensorflow.py
@@ -1,6 +1,6 @@
-from sympy.printing.lambdarepr import TensorflowPrinter
-from sympy.printing.lambdarepr import tensorflow_code
-from sympy import (eye, symbols, MatrixSymbol, Symbol)
+from sympy.printing.tensorflow import TensorflowPrinter
+from sympy.printing.tensorflow import tensorflow_code
+from sympy import (eye, symbols, MatrixSymbol, Symbol, Matrix, symbols, sin, exp)
 from sympy.codegen.array_utils import (CodegenArrayContraction,
         CodegenArrayTensorProduct, CodegenArrayElementwiseAdd,
         CodegenArrayPermuteDims, CodegenArrayDiagonal)
@@ -16,12 +16,17 @@ N = MatrixSymbol("N", 3, 3)
 P = MatrixSymbol("P", 3, 3)
 Q = MatrixSymbol("Q", 3, 3)
 
+x, y, z, t = symbols("x y z t")
+
 
 def test_tensorflow_matrix():
     if not tf:
         skip("TensorFlow not installed")
 
     assert tensorflow_code(eye(3)) == "tensorflow.constant([[1, 0, 0], [0, 1, 0], [0, 0, 1]])"
+
+    expr = Matrix([[x, sin(y)], [exp(z), -t]])
+    assert tensorflow_code(expr) == "tensorflow.Variable([[x, tensorflow.sin(y)], [tensorflow.exp(z), -t]])"
 
     expr = M
     assert tensorflow_code(expr) == "M"
@@ -31,6 +36,8 @@ def test_tensorflow_matrix():
     assert tensorflow_code(expr) == "tensorflow.matmul(M, N)"
     expr = M*N*P*Q
     assert tensorflow_code(expr) == "tensorflow.matmul(tensorflow.matmul(tensorflow.matmul(M, N), P), Q)"
+    expr = M**3
+    assert tensorflow_code(expr) == "tensorflow.matmul(tensorflow.matmul(M, M), M)"
 
 
 def test_codegen_einsum():
@@ -115,3 +122,15 @@ def test_codegen_extra():
     y = session.run(f(ma, mb))
     c = session.run(tf.einsum("ab,bc->acb", ma, mb))
     assert (y == c).all()
+
+
+def test_MatrixElement_printing():
+    A = MatrixSymbol("A", 1, 3)
+    B = MatrixSymbol("B", 1, 3)
+    C = MatrixSymbol("C", 1, 3)
+
+    assert tensorflow_code(A[0, 0]) == "A[0, 0]"
+    assert tensorflow_code(3 * A[0, 0]) == "3*A[0, 0]"
+
+    F = C[0, 0].subs(C, A - B)
+    assert tensorflow_code(F) == "((-1)*B + A)[0, 0]"
